# LibSoftGPU+LibGfx: Improve triangle rendering quality