We used to convert every triangle's window coordinates to `int` screen coordinates, losing a bit too much precision in the process. Two triangles could partially share an edge, but because the vertex positions were casted to `int`s some gaps could occur in between what would otherwise be perfectly colinear edges. This PR converts the logic to use `float` precision instead. This fixes the triangle edge artifacts that were visible in GLQuake and improves the ice reflections in Tux Racer.

Also: some `Gfx::Vector` improvements :^)

Some super subjective benchmarks on my machine, eyeballing the FPS counter:

**GLQuake benchmark**
Start a new game, turn around to face the wall.

| Build | `master` | PR |
| ----- | -------- | -- |
| i686 | 2 FPS | 3 FPS |
| i686 + SSE1+2 | 29 FPS | 30 FPS |
| x86_64 | 37 FPS | 39 FPS |

**Tux Racer benchmark**
Start a practice run on the default level, measure during intro waddle.

| Build | `master` | PR |
| ----- | -------- | -- |
| i686 | 2 FPS | 2 FPS |
| i686 + SSE1+2 | 17 FPS | 17 FPS |
| x86_64 | 21 FPS | 21 FPS |

**3DFileViewer benchmark**
Open application and measure immediately.

| Build | `master` | PR |
| ----- | -------- | -- |
| i686 | 16 FPS | 16 FPS |
| i686 + SSE1+2 | 41 FPS | 44 FPS |
| x86_64 | 48 FPS | 47 FPS |

GLquake before: (white pixels to the left of the gun)
![image](https://user-images.githubusercontent.com/3210731/156937959-ea6430b8-0965-4bd8-ba3a-c204bc826a25.png)

After: (no white pixels to be seen)
![image](https://user-images.githubusercontent.com/3210731/156937967-a843df87-824d-4c75-ac10-80b733505b47.png)