@GMTA 2022-03-06
@ldm5180 could you review the `VectorN.h` changes in this PR please?